Rename modules - Group thresholding methods in data
